refactor: reorganize file
